fix: resolve Send/Sync violations and implement safe CoreML integration
- Worker A: Council compile unblock (Send/Sync violations)
  - Created thread-confined CoreML handles with !Send/!Sync semantics
  - Implemented channel-based communication between async coordination and inference threads
  - Added ModelClient for safe async communication with inference worker

- Worker B: Safe FFI façade for CoreML operations
  - Implemented ModelRef(u64) opaque identifiers that are Send/Sync
  - Created thread-local registry system mapping ModelRef to CoreMlHandle
  - Added proper resource cleanup and memory safety guarantees

- Worker C: Legacy type reference cleanup
  - Updated all model implementations (Mistral, Whisper, Diffusion) to use new safe types
  - Replaced ModelHandle references with ModelRef across apple-silicon crate
  - Fixed import paths and module declarations

- Worker D: Council integration and telemetry cleanup
  - Integrated ModelClient into MistralJudge with channel-based communication
  - Removed incompatible telemetry API calls (emit_event, record_metric)
  - Added local type definitions for InferenceRequest/InferenceResult

Safety improvements:
- Zero Send/Sync violations - raw pointers properly quarantined to dedicated threads
- Thread-safe model references that can cross async boundaries
- Memory leak prevention with proper Drop implementations
- Comprehensive error handling and resource cleanup

Both council and apple-silicon crates now compile successfully with zero errors.

+++ iterations/v3/parallel-workers/src/worker/context.rs
@@ -317,10 +317,10 @@ impl ExecutionSandbox {
         // Execute with timeout
         match tokio::time::timeout(self.resource_limits.time_limit, cmd.output()).await {
             Ok(result) => result.map_err(|e| WorkerError::ExecutionTimeout {
-                timeout_secs: self.resource_limits.time_limit.as_secs() as u64,
+                timeout_secs: self.resource_limits.time_limit.as_secs(),
             }),
             Err(_) => Err(WorkerError::ExecutionTimeout {
-                timeout_secs: self.resource_limits.time_limit.as_secs() as u64,
+                timeout_secs: self.resource_limits.time_limit.as_secs(),
             }),
         }
     }
